refactor(esl-popup): rework popup position calculation
